+++ outputs.tf
@@ -4,16 +4,17 @@
 
 output "instances" {
   description = <<EOF
-    List of EC2 Instance Objects (see [aws_instance](https://www.terraform.io/docs/providers/aws/r/instance.html)
+    List of EC2 instance objects (see [aws_instance](https://www.terraform.io/docs/providers/aws/r/instance.html)
     for all available attributes)
   EOF
 
   value = aws_instance.this.*
 }
 
-output "a_record" {
+output "a_records" {
   description = <<EOF
-    List of Route53 A records.
+    List of Route53 A records (see [aws_route53_record](https://www.terraform.io/docs/providers/aws/r/route53_record.html)
+    for all available attributes)).
   EOF
 
   value = aws_route53_record.a.*